requirements.txt
boto3==1.18.47
mysql-connector-python==8.0.28
